requirements.txt
django==1.8.7
django-flat-theme==1.1.3
Pillow==3.0.0
django-mptt==0.7.4
django_mptt_admin==0.3.1
django-ckeditor==5.0.2
django-image-cropping==1.0.2
easy_thumbnails==2.0.1
django-sortedm2m==1.1.0
django-tabbed-admin==0.0.5
django-debug-toolbar==1.4
